fix: Relocated files based on domains. Removed unused files and configs.
