+++ src/app/presentation/components/sidebar/sidebar.component.spec.ts
@@ -7,8 +7,8 @@ import { MatListModule } from '@angular/material/list';
 import { MatIconModule } from '@angular/material/icon';
 import { CommonModule } from '@angular/common';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
-import { LoginService } from './../../../core/services/login/login.service';
 import { RouteData } from '../../interfaces/route-data.interface';
+import { LoginService } from '../../../data/auth/services/login/login.service';
 
 describe('SidebarComponent', () => {
   let component: SidebarComponent;
